Improve test_llgroup_size_limit. Fix: cast to unsigned short only after the division.

+++ pypy/translator/c/src/llgroup.h
@@ -2,7 +2,7 @@
 #define _PYPY_LL_GROUP_H_
 
 #define GROUP_MEMBER_OFFSET(grouptype, membername)  \
-    ((unsigned short)(&(((grouptype*)NULL))->membername) / sizeof(long))
+  ((unsigned short)(((int)&((grouptype*)NULL)->membername) / sizeof(long)))
 
 #define _OP_GET_GROUP_MEMBER(groupptr, compactoffset)  \
   (((char*)groupptr) + ((long)compactoffset)*sizeof(long))
